Modify when council member is informed
Currently potential council member is asked about his willingness
to participate before a formal vote. Although rarerly, a vote on
a potential new member can vote and thus make the whole interaction
a little degrading. Informing the potential member after a successful
vote remove that possibility and still gives the opportunity for the
potential member to decline the invitation.