Clarify usage of documents that may contains images of the trademark.
Derivated from a private request asking is Jupyter Can be use in a
Certification program, it looks like the requested is concern that not
all case are covered by this section.

My understanding is that it's ok to _use_ jupyter in your certification
program and provide screenshots, video of it. I don't believe this about
becoming Certified _in_ Jupyter, which I'm not sure what that would mean
and is another question.