Add Project Jupyter organizations to README
As we've grown, projects have split off into new orgs. I couldn't find a list of all the current GitHub orgs that contain Project Jupyter repos. I'm adding a list here, but I'm not sure if I am missing any orgs. 

Once we merge this, it would be helpful to have this list added somewhere in the website.