Clarify that I'm also stepping down as BDFL
This has been always the intent and stated many times verbally, but @jasongrout just realized we'd not put it in writing in this document. It's a good idea to document it explicitly in case anyone wonders.

The BDFL idea was one we took from Python back very early in the IPython days, but this entire new model is meant to replace centralized, personal control with an open community approach. So we want to be explicit about it.