trademarks: allow base domains where
The goal of the domain rules is to avoid confusion about whether a service is affiliated with the Jupyter project. Make it explicit that such confusion can still be resolved in the base domain itself.

For secure deployment reasons like cookie tossing, we should be _recommending_ hosted jupyter deployments be on their own domain (for the same reasons github has githubusercontent.com and github.io). The trademark policy shouldn't make it harder for folks to follow best practices.