Half the number of yearly Distinguished Contributor
We had more spot than candidates for a few years,
this is in my opinion problematic.

Among other,

 - It decreases the recognition of the award
 - It forces us to change voting procedure

I'm open to decrease to less than 5, but I think
this is  a first step.

Note that I suggested this a couple of time, but was told this
was too close to next elections, as this time the election are just finished,
I hope this is early enough.