Generate the leadership directory from structured data
And allow to include the leadership directory in multiple documents.

There are a few visual changes to uniformize sections headers.
Like terms between parentheses.

The column order of SSC and governing board. and the

> Former Standing Committees and Working Groups
>    [Name former groups]
>    [Name former groups]
>     ...

Which is now:

> Former x
>   Former User 1
>   Former User 2

Also readed myself as former trademark
subcommittee members as I was gone, not sure why.

The bulk of the extration of data into a .yml file
was done using an LLM.